add grammar

+++ legend-engine-xt-relationalStore-grammar/src/main/java/org/finos/legend/engine/language/pure/compiler/toPureGraph/DatasourceSpecificationBuilder.java
@@ -144,7 +144,7 @@ public class DatasourceSpecificationBuilder implements DatasourceSpecificationVi
             spannerSpec._projectId(spannerDatasourceSpecification.projectId);
             spannerSpec._instanceId(spannerDatasourceSpecification.instanceId);
             spannerSpec._proxyHost(spannerDatasourceSpecification.proxyHost);
-            spannerSpec._proxyPort(spannerDatasourceSpecification.proxyPort);
+            spannerSpec._proxyPort(spannerDatasourceSpecification.proxyPort.longValue());
             spannerSpec._databaseId(spannerDatasourceSpecification.databaseId);
             return spannerSpec;
         }
